refactor(nat): Move NatIp trait def/implementation to own submodule
In preparation for more work on stateful NAT, namely for the IP/port
allocator, move the definition and implmementation of the NatIp trait
(representing either IPv4 or IPv6 addresses, for NAT) to its own
submodule, to make future changes to this trait easier to follow.

Also make Clippy complain again on unused imports for the stateful NAT
module (currently no report).

No functional change.
